requirements.txt
fastapi
uvicorn
openai
python-dotenv
pandas
pdfplumber
docx2txt
python-multipart
